# public/swagger/swagger.yml
openapi: 3.0.0
info:
  title: 'TiSaude API'
  version: '0.1'
paths:
...
  /api/procedures:
    get:
      tags:
        - Procedimentos
      summary: 'Lista de Procedimentos'
      description: 'Lista de Procedimentos'
      operationId: 7d6ba815589b5f6a122b0c3362e8b177
      responses:
        '200':
          description: 'Lista de Procedimentos'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Procedimentos
      summary: 'Cria um novo Procedimento'
      description: 'Cria um novo Procedimento'
      operationId: 65ae2b00841b4a1ab0896811636e8065
      parameters:
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Cria um novo Procedimento'
      security:
        -
          bearerAuth: []
  '/api/procedures/{id}':
    get:
      tags:
        - Procedimentos
      summary: 'Exibe um procedimento'
      description: 'Exibe um procedimento'
      operationId: 7d7cd20f72364dc472c949ad5465383b
      parameters:
        -
          name: id
          in: path
          description: 'ID do Procedimento'
          required: true
      responses:
        '200':
          description: 'Exibe um procedimento'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Procedimentos
      summary: 'Atualiza um procedimento'
      description: 'Atualiza um procedimento'
      operationId: b02480753e102fe3bca0473abf88b626
      parameters:
        -
          name: id
          in: path
          description: 'ID do Procedimento'
          required: true
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Atualiza um procedimento'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Procedimentos
      summary: 'Deleta um procedimento'
      description: 'Deleta um procedimento'
      operationId: a9c66777a1bdff4ce6549c220a927bfa
      parameters:
        -
          name: id
          in: path
          description: 'ID do Procedimento'
          required: true
      responses:
        '200':
          description: 'Deleta um procedimento'
      security:
        -
          bearerAuth: []
  /api/specialties:
    get:
      tags:
        - Especialidades
      summary: 'Lista as especialidades'
      description: 'Lista as especialidades'
      operationId: 67515fb7ae55ba00df38f3069860eba8
      responses:
        '200':
          description: 'Lista as especialidades'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Especialidades
      summary: 'Cria uma nova especialidades'
      description: 'Cria uma nova especialidades'
      operationId: 6ffffb6d94d102c82f2081765ba7bdae
      parameters:
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Cria uma nova especialidades'
      security:
        -
          bearerAuth: []
  '/api/specialties/{id}':
    get:
      tags:
        - Especialidades
      summary: 'Lista uma especialidade'
      description: 'Lista uma especialidade'
      operationId: b2731d1377c994d3814193cd2edcc834
      parameters:
        -
          name: id
          in: path
          description: 'ID da Especialidade'
          required: true
      responses:
        '200':
          description: 'Lista uma especialidade'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Especialidades
      summary: 'Atualiza uma especialidades'
      description: 'Atualiza uma especialidades'
      operationId: 00783a2640e0f8ddb81ccfeb1cecd2e2
      parameters:
        -
          name: id
          in: path
          description: 'ID da Especialidade'
          required: true
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Atualiza uma especialidades'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Especialidades
      summary: 'Deleta uma especialidades'
      description: 'Deleta uma especialidades'
      operationId: 129546f791893572810b5f9006c6de58
      parameters:
        -
          name: id
          in: path
          description: 'ID da Especialidade'
          required: true
      responses:
        '200':
          description: 'Deleta uma especialidades'
      security:
        -
          bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header